requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
yt-dlp
google-generativeai
python-dotenv
